// test-app/src/generated-types.ts
export type Maybe<T> = T | null | undefined;
export type InputMaybe<T> = T | null | undefined;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
export { ResolverFn };
export type RequireFields<T, K extends keyof T> = Omit<T, K> & { [P in K]-?: NonNullable<T[P]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};

export type Mutation = {
  addPerson: Maybe<Person>;
  deletePerson: Maybe<Person>;
  updateName: Maybe<Person>;
};


export type MutationAddPersonArgs = {
  input: PersonInput;
};


export type MutationDeletePersonArgs = {
  id: Scalars['ID'];
};


export type MutationUpdateNameArgs = {
  id: Scalars['ID'];
  name: Scalars['String'];
};

export type Person = {
  id: Scalars['String'];
  name: Scalars['String'];
};

export type PersonInput = {
  name: Scalars['String'];
};

export type ProcessedPerson = {
  id: Scalars['String'];
  name: Scalars['String'];
  score: Scalars['Int'];
};

export type Query = {
  getPerson: Maybe<ProcessedPerson>;
};


export type QueryGetPersonArgs = {
  id: Scalars['String'];
};



export type ResolverTypeWrapper<T> = Promise<T> | T;

export type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> = ResolverFn<TResult, TParent, TContext, TArgs>;
// ...
export type ResolversTypes = {
  Boolean: ResolverTypeWrapper<Scalars['Boolean']>;
  ID: ResolverTypeWrapper<Scalars['ID']>;
  Int: ResolverTypeWrapper<Scalars['Int']>;
  Mutation: ResolverTypeWrapper<{}>;
  Person: ResolverTypeWrapper<Person>;
  PersonInput: PersonInput;
  ProcessedPerson: ResolverTypeWrapper<ProcessedPerson>;
  Query: ResolverTypeWrapper<{}>;
  String: ResolverTypeWrapper<Scalars['String']>;
};

/** Mapping between all available schema types and the resolvers parents */
export type ResolversParentTypes = {
  Boolean: Scalars['Boolean'];
  ID: Scalars['ID'];
  Int: Scalars['Int'];
  Mutation: {};
  Person: Person;
  PersonInput: PersonInput;
  ProcessedPerson: ProcessedPerson;
  Query: {};
  String: Scalars['String'];
};

export type MutationResolvers<ContextType = any, ParentType extends ResolversParentTypes['Mutation'] = ResolversParentTypes['Mutation']> = {
  addPerson: Resolver<Maybe<ResolversTypes['Person']>, ParentType, ContextType, RequireFields<MutationAddPersonArgs, 'input'>>;
  deletePerson: Resolver<Maybe<ResolversTypes['Person']>, ParentType, ContextType, RequireFields<MutationDeletePersonArgs, 'id'>>;
  updateName: Resolver<Maybe<ResolversTypes['Person']>, ParentType, ContextType, RequireFields<MutationUpdateNameArgs, 'id' | 'name'>>;
};
// ...
export type QueryResolvers<ContextType = any, ParentType extends ResolversParentTypes['Query'] = ResolversParentTypes['Query']> = {
  getPerson: Resolver<Maybe<ResolversTypes['ProcessedPerson']>, ParentType, ContextType, RequireFields<QueryGetPersonArgs, 'id'>>;
};